+++ examples/config.php
@@ -1,4 +1,7 @@
 <?php
+
+use Moudarir\File\Enum\MimeType;
+
 return [
     'root_path' => dirname(__DIR__) . DIRECTORY_SEPARATOR,
     'urls' => [
@@ -11,33 +14,33 @@ return [
             'mov' => [
                 'basename' => 'video.mov',
                 'filename' => 'test video.mov',
-                'mime' => 'video/quicktime',
+                'mime' => MimeType::MOV,
             ],
             'pdf' => [
                 'basename' => 'document.pdf',
                 'filename' => 'example.pdf',
-                'mime' => 'application/pdf',
+                'mime' => MimeType::PDF,
             ],
             'bin' => [
                 'basename' => 'resource.bin',
                 'filename' => 'resource.bin',
-                'mime' => 'application/octet-stream',
+                'mime' => MimeType::OCTET_STREAM,
             ],
             'txt' => [
                 'basename' => 'text.txt',
                 'filename' => 'text.txt',
-                'mime' => 'text/plain',
+                'mime' => MimeType::TEXT_PLAIN,
             ],
         ],
         'data' => [
             'content' => "lorem ipsum dolor sit amet lorem ipsum dolor sit amet lorem ipsum dolor sit amet",
             'filename' => 'data.txt',
-            'mime' => 'text/plain',
+            'mime' => MimeType::TEXT_PLAIN,
             'etag' => '9faa0b435bc1eec824d2e54ca30deb04',
         ],
     ],
     'multipart' => [
         'boundary' => '3d6b6a416f9b5d3b',
-        'mime' => 'text/plain',
+        'mime' => MimeType::TEXT_PLAIN,
     ],
 ];